improve `umi build --watch`, support watch and rebuild tmp files

+++ packages/umi-build-dev/src/plugins/commands/build/index.js
@@ -14,7 +14,8 @@ export default function(api) {
       webpack: true,
       description: 'building for production',
     },
-    () => {
+    args => {
+      const watch = args.w || args.watch;
       notify.onBuildStart({ name: 'umi', version: 2 });
 
       const RoutesManager = getRouteManager(service);
@@ -24,11 +25,24 @@ export default function(api) {
         process.env.NODE_ENV = 'production';
         service.applyPlugins('onStart');
         service._applyPluginsAsync('onStartAsync').then(() => {
+          service.rebuildTmpFiles = () => {
+            filesGenerator.rebuild();
+          };
+          service.rebuildHTML = () => {
+            service.applyPlugins('onHTMLRebuild');
+          };
+
           const filesGenerator = getFilesGenerator(service, {
             RoutesManager,
             mountElementId: config.mountElementId,
           });
           filesGenerator.generate();
+
+          function startWatch() {
+            filesGenerator.watch();
+            service.userConfig.setConfig(service.config);
+            service.userConfig.watchWithDevServer();
+          }
 
           if (process.env.HTML !== 'none') {
             const HtmlGeneratorPlugin = require('../getHtmlGeneratorPlugin').default(service);
@@ -40,13 +54,17 @@ export default function(api) {
           service._applyPluginsAsync('beforeBuildCompileAsync').then(() => {
             require('af-webpack/build').default({
               cwd,
+              watch,
               webpackConfig: [
                 service.webpackConfig,
                 ...(service.ssrWebpackConfig ? [service.ssrWebpackConfig] : []),
               ],
               onSuccess({ stats }) {
                 debug('Build success');
-                if (process.env.RM_TMPDIR !== 'none') {
+                if (watch) {
+                  startWatch();
+                }
+                if (process.env.RM_TMPDIR !== 'none' && !watch) {
                   debug(`Clean tmp dir ${service.paths.tmpDirPath}`);
                   rimraf.sync(paths.absTmpDirPath);
                 }
